Реализует BoardController
