+++ src/main.js
@@ -2,11 +2,10 @@ import {TASK_COUNT} from "./common/consts";
 import {render} from "./common/utils/render";
 import {generateFilters} from './mock/filters.js';
 import {generateTasks} from "./mock/tasks";
-import {renderBoard} from "./components/board/utils/render-board";
-import {renderTask} from './components/task/utils/render-task';
 import MenuComponent from "./components/menu/menu";
 import FiltersComponent from "./components/filter/filters";
 import BoardComponent from "./components/board/board";
+import BoardController from './controllers/board';
 
 const siteMain = document.querySelector(`.main`);
 const siteHeader = siteMain.querySelector(`.main__control`);
@@ -16,11 +15,13 @@ const tasks = generateTasks(TASK_COUNT);
 
 const init = () => {
   const boardComponent = new BoardComponent(tasks);
+  const boardController = new BoardController(boardComponent);
 
   render(siteHeader, new MenuComponent());
   render(siteMain, new FiltersComponent(filters));
   render(siteMain, boardComponent);
-  renderBoard(boardComponent, tasks, renderTask);
+
+  boardController.render(tasks);
 };
 
 init();
